requirements.txt
# alembic==1.0.0.dev0
# apipkg==1.5
# # apturl==0.5.2
# asgiref==3.2.10
# asn1crypto==0.24.0
# attrs==20.3.0
# Babel==2.6.0
# bcrypt==3.1.6
blinker==1.4
# Brlapi==0.6.7
# certifi==2018.8.24
# chardet==3.0.4
# Click==7.0
# colorama==0.3.7
# command-not-found==0.3
# cryptography==2.3
# # cupshelpers==1.0
# # defer==1.0.6
# defusedxml==0.6.0
# distro==1.3.0
# distro-info===0.21ubuntu2
Django==3.1
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
# entrypoints==0.3
# execnet==1.7.1
# Flask==1.0.2
# Flask-BabelEx==0.9.3
# Flask-Compress==1.4.0
# Flask-Gravatar==0.4.2
# Flask-Login==0.4.1
# Flask-Mail==0.9.1
# Flask-Migrate==2.1.1
# Flask-Paranoid==0.2.0
# Flask-Principal==0.4.0
# Flask-Security==1.7.5
# Flask-SQLAlchemy==2.1
# Flask-WTF==0.14.2
# httplib2==0.11.3
# idna==2.6
# importlib-metadata==3.4.0
# iniconfig==1.1.1
# itsdangerous==0.24
# Jinja2==2.10
# keyring==17.1.1
# keyrings.alt==3.1.1
# language-selector==0.1
# launchpadlib==1.10.6
# lazr.restfulclient==0.14.2
# lazr.uri==1.0.3
# louis==3.8.0
# macaroonbakery==1.2.1
# Mako==1.0.7
# MarkupSafe==1.1.0
mysqlclient==2.0.3
# netifaces==0.10.4
oauthlib==3.1.0
# olefile==0.46
# packaging==20.8
# paramiko==2.4.2
# passlib==1.7.1
# pexpect==4.6.0
Pillow==5.4.1
# pluggy==0.13.1
# protobuf==3.6.1
# psutil==5.5.1
# psycopg2==2.8.4
# py==1.10.0
# pyasn1==0.4.2
# pycairo==1.16.2
# pycrypto==2.6.1
# pycups==1.9.73
# PyGObject==3.32.0
# pyinotify==0.9.6
# PyJWT==1.7.0
# pymacaroons==0.13.0
PyMySQL==1.0.2
# PyNaCl==1.3.0
# pyOpenSSL==19.0.0
# pyparsing==2.4.7
# pyRFC3339==1.1
pytest==6.2.1
# pytest-cache==1.0
# python-apt===1.8.5-ubuntu0.3
# python-dateutil==2.7.3
# python-debian==0.1.34
# python3-openid==3.2.0
# pytz==2018.9
# pyxdg==0.25
# PyYAML==3.13
# reportlab==3.5.18
# requests==2.21.0
# requests-oauthlib==1.3.0
# requests-unixsocket==0.1.5
# SecretStorage==2.3.1
selenium==3.141.0
# simplejson==3.16.0
# six==1.12.0
# SQLAlchemy==1.2.15
# sqlparse==0.2.4
# sshtunnel==0.1.4
# system-service==0.3
# systemd-python==234
# toml==0.10.2
# typing-extensions==3.7.4.3
# ubuntu-advantage-tools==19.2
# ubuntu-drivers-common==0.0.0
# ufw==0.36
# unattended-upgrades==0.1
# urllib3==1.24.1
# usb-creator==0.3.3
# wadllib==1.3.3
# Werkzeug==0.16.0
webdriver-manager==3.4.2
# WTForms==2.2.1
# xkit==0.0.0
# zipp==3.4.0
